Coverage: move _compile method to fixtures.js
