+++ lib/compiler.js
@@ -75,22 +75,3 @@ Compiler.prototype.compile = function compile(code, options) {
 
   return exports;
 };
-
-Compiler.prototype._compile = function _compile(code, options) {
-  if (!options) options = {};
-
-  // It is fine to compile without templates at first
-  if (!code)
-    code = function() {};
-
-  var engineName = options.engine || 'BEMHTML';
-
-  var Engine = require('./' + engineName.toLowerCase());
-  var api = new Engine(options);
-
-  var template = {};
-  api.compile(code);
-  api.exportApply(template);
-
-  return template;
-};
